Implementados services, corrigidos erros de tipos, corrigidos erros de nomes de métodos comunicando entre os services e controllers

+++ app/src/main/java/com/ecommerce/app/controller/product/ProductController.java
@@ -22,7 +22,7 @@ public class ProductController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<ProductDTO> getProductById(@PathVariable int id){
+    public ResponseEntity<ProductDTO> getProductById(@PathVariable Long id){
         ProductDTO productDTO = productService.getProductById(id);
 
         return productDTO != null ? ResponseEntity.ok(productDTO) : ResponseEntity.notFound().build();
@@ -34,14 +34,14 @@ public class ProductController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<ProductDTO> updateProduct(@PathVariable int id, @RequestBody ProductDTO productDTO){
+    public ResponseEntity<ProductDTO> updateProduct(@PathVariable Long id, @RequestBody ProductDTO productDTO){
         ProductDTO updateProduct = productService.updateProduct(id, productDTO);
 
         return updateProduct != null ? ResponseEntity.ok(updateProduct) : ResponseEntity.notFound().build();
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteProduct(@PathVariable int id){
+    public ResponseEntity<Void> deleteProduct(@PathVariable Long id){
         productService.deleteProduct(id);
         return ResponseEntity.noContent().build();
     }
